Add `noMultiTouch` option to prevent reacting to multi touch events
This adds a new setting noMultiTouch and makes it the default.
With `noMultiTouch` set to true snap.js won't react to multi
finger gestures. This should be the default as it's quite unlikely
to be what the user expects. It helps to prevent issues
with other gestures like pinch zooming.